Updated BCI2000 signal type to float32 instead of an integer type which was effectively reducing the resolution of the signal significantly.
git-svn-id: http://www.bci2000.org/svn/trunk@2112 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalSource/AmpServerPro/AmpServerProADC.cpp
@@ -395,7 +395,7 @@ void AmpServerProADC::Preflight( const SignalProperties&, SignalProperties& outS
   
   //samples are always 4-byte floats
   outSignalProperties = SignalProperties(
-       Parameter( "SourceCh" ), Parameter( "SampleBlockSize" ), SignalType::int32 );
+       Parameter( "SourceCh" ), Parameter( "SampleBlockSize" ), SignalType::float32);
   nScanRes = sscanf(Parameter( "SampleBlockSize" ).c_str(), "%d", &nBlockSize);
   if(nScanRes <= 0 || nScanRes == EOF)
   {
